working version

+++ TextAnalyzer_CommandLine/src/functions/IDF_LogSmooth.java
@@ -9,6 +9,6 @@ public class IDF_LogSmooth implements FunctionIDF {
 	@Override
 	public double calculate(String term, List<WeightedMap> inputCorpus) {
 		long count = inputCorpus.stream().filter(val -> val.containsKey(term) && val.get(term) != 0).count();
-		return Math.log(inputCorpus.size() / count);
+		return Math.log((double)inputCorpus.size() / (double)count);
 	}
 }
